refactor(enums): consolidate and organize enums for STI test processes into a dedicated file refactor(menstrual-predictions): update cycle ordering to use SortOrder enum for consistency refactor(app): update environment configuration and disable logging for production
